refactor global database tests
